requirements.txt
colorama==0.4.6
greenlet==3.3.2
iniconfig==2.3.0
packaging==26.0
playwright==1.58.0
pluggy==1.6.0
pyee==13.0.1
Pygments==2.19.2
pytest==9.0.2
typing_extensions==4.15.0
allure-pytest==2.15.3

